backend/requirements.txt
aniso8601==9.0.1
Authlib==1.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
Flask==2.2.2
Flask-RESTful==0.3.9
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycparser==2.21
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
urllib3==1.26.14
Werkzeug==2.2.2
psycopg2-binary==2.9.5
aniso8601==9.0.1
Authlib==1.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
pycparser==2.21
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
urllib3==1.26.14
Werkzeug==2.2.2
zipp==3.15.0
